ditch swift 5.7/5.8
