+++ Macros/UnstableEnumMacro/EntryPoint.swift
@@ -4,6 +4,6 @@ import SwiftCompilerPlugin
 @main
 struct EntryPoint: CompilerPlugin {
     let providingMacros: [any Macro.Type] = [
-        UnstableEnumMacro.self
+        UnstableEnum.self
     ]
 }
