+++ Sources/DiscordModels/Types/Gateway.swift
@@ -509,7 +509,6 @@ public struct Gateway: Sendable, Codable {
     }
 
     /// https://discord.com/developers/docs/topics/gateway#gateway-intents
-#if swift(>=5.9) && $Macros
     @UnstableEnum<UInt>
     public enum Intent: Sendable, Codable, CaseIterable {
         case guilds // 0
@@ -532,29 +531,6 @@ public struct Gateway: Sendable, Codable {
         case autoModerationConfiguration // 20
         case autoModerationExecution // 21
     }
-#else
-    public enum Intent: UInt, Sendable, Codable, CaseIterable {
-        case guilds = 0
-        case guildMembers = 1
-        case guildModeration = 2
-        case guildEmojisAndStickers = 3
-        case guildIntegrations = 4
-        case guildWebhooks = 5
-        case guildInvites = 6
-        case guildVoiceStates = 7
-        case guildPresences = 8
-        case guildMessages = 9
-        case guildMessageReactions = 10
-        case guildMessageTyping = 11
-        case directMessages = 12
-        case directMessageReactions = 13
-        case directMessageTyping = 14
-        case messageContent = 15
-        case guildScheduledEvents = 16
-        case autoModerationConfiguration = 20
-        case autoModerationExecution = 21
-    }
-#endif
 
     /// https://discord.com/developers/docs/topics/gateway-events#resume-resume-structure
     public struct Resume: Sendable, Codable {
@@ -570,7 +546,6 @@ public struct Gateway: Sendable, Codable {
     }
     
     /// https://discord.com/developers/docs/topics/gateway-events#update-presence-status-types
-#if swift(>=5.9) && $Macros
     @UnstableEnum<String>
     public enum Status: Sendable, Codable {
         case online // "online"
@@ -579,15 +554,6 @@ public struct Gateway: Sendable, Codable {
         case offline // "offline"
         case invisible // "invisible"
     }
-#else
-    public enum Status: String, Sendable, Codable, ToleratesStringDecodeMarker {
-        case online = "online"
-        case doNotDisturb = "dnd"
-        case afk = "idle"
-        case offline = "offline"
-        case invisible = "invisible"
-    }
-#endif
 
     /// https://discord.com/developers/docs/topics/gateway-events#hello-hello-structure
     public struct Hello: Sendable, Codable {
@@ -931,18 +897,11 @@ public struct Gateway: Sendable, Codable {
     public struct InviteCreate: Sendable, Codable {
         
         /// https://discord.com/developers/docs/resources/invite#invite-object-invite-target-types
-#if swift(>=5.9) && $Macros
         @UnstableEnum<Int>
         public enum TargetKind: Sendable, Codable {
             case stream // 1
             case embeddedApplication // 2
         }
-#else
-        public enum TargetKind: Int, Sendable, Codable, ToleratesIntDecodeMarker {
-            case stream = 1
-            case embeddedApplication = 2
-        }
-#endif
 
         public var channel_id: ChannelSnowflake
         public var code: String
@@ -1085,18 +1044,11 @@ public struct Gateway: Sendable, Codable {
         public var guild_id: GuildSnowflake?
     }
 
-#if swift(>=5.9) && $Macros
     @UnstableEnum<Int>
     public enum ReactionKind: Sendable, Codable {
         case normal // 0
         case `super` // 1
     }
-#else
-    public enum ReactionKind: Int, Sendable, Codable {
-        case normal = 0
-        case `super` = 1
-    }
-#endif
 
     /// https://discord.com/developers/docs/topics/gateway-events#message-reaction-add-message-reaction-add-event-fields
     public struct MessageReactionAdd: Sendable, Codable {
@@ -1185,7 +1137,6 @@ public struct Gateway: Sendable, Codable {
     public struct Activity: Sendable, Codable {
         
         /// https://discord.com/developers/docs/topics/gateway-events#activity-object-activity-types
-#if swift(>=5.9) && $Macros
         @UnstableEnum<Int>
         public enum Kind: Sendable, Codable {
             case game // 0
@@ -1195,16 +1146,6 @@ public struct Gateway: Sendable, Codable {
             case custom // 4
             case competing // 5
         }
-#else
-        public enum Kind: Int, Sendable, Codable, ToleratesIntDecodeMarker {
-            case game = 0
-            case streaming = 1
-            case listening = 2
-            case watching = 3
-            case custom = 4
-            case competing = 5
-        }
-#endif
 
         /// https://discord.com/developers/docs/topics/gateway-events#activity-object-activity-timestamps
         public struct Timestamps: Sendable, Codable {
@@ -1270,7 +1211,6 @@ public struct Gateway: Sendable, Codable {
         }
 
         /// https://discord.com/developers/docs/topics/gateway-events#activity-object-activity-flags
-#if swift(>=5.9) && $Macros
         @UnstableEnum<UInt>
         public enum Flag: Sendable {
             case instance // 0
@@ -1283,19 +1223,6 @@ public struct Gateway: Sendable, Codable {
             case partyPrivacyVoiceChannel // 7
             case embedded // 8
         }
-#else
-        public enum Flag: UInt, Sendable {
-            case instance = 0
-            case join = 1
-            case spectate = 2
-            case joinRequest = 3
-            case sync = 4
-            case play = 5
-            case partyPrivacyFriends = 6
-            case partyPrivacyVoiceChannel = 7
-            case embedded = 8
-        }
-#endif
 
         /// https://discord.com/developers/docs/topics/gateway-events#activity-object-activity-buttons
         public struct Button: Sendable, Codable {
@@ -1416,12 +1343,10 @@ extension Gateway.Intent {
         case .guildScheduledEvents: return false
         case .autoModerationConfiguration: return false
         case .autoModerationExecution: return false
-#if swift(>=5.9) && $Macros
-            /// Unknown cases are considered privileged just to safe than sorry
+            /// Unknown cases are considered privileged just to be safe than sorry
         case .unknown: return true
         case .__DO_NOT_USE_THIS_CASE:
             fatalError("If the case name wasn't already clear enough: This case MUST NOT be used under any circumstances")
-#endif
         }
     }
 }
