+++ Sources/DiscordModels/Types/Permission.swift
@@ -1,6 +1,5 @@
 
 /// https://discord.com/developers/docs/topics/permissions#permissions-bitwise-permission-flags
-#if swift(>=5.9) && $Macros
 @UnstableEnum<UInt>
 public enum Permission: Sendable, Hashable, Codable {
     case createInstantInvite // 0
@@ -49,55 +48,6 @@ public enum Permission: Sendable, Hashable, Codable {
     case useExternalSounds // 45
     case sendVoiceMessages // 46
 }
-#else
-public enum Permission: UInt, Sendable, Hashable, Codable {
-    case createInstantInvite = 0
-    case kickMembers = 1
-    case banMembers = 2
-    case administrator = 3
-    case manageChannels = 4
-    case manageGuild = 5
-    case addReactions = 6
-    case viewAuditLog = 7
-    case prioritySpeaker = 8
-    case stream = 9
-    case viewChannel = 10
-    case sendMessages = 11
-    case sendTtsMessages = 12
-    case manageMessages = 13
-    case embedLinks = 14
-    case attachFiles = 15
-    case readMessageHistory = 16
-    case mentionEveryone = 17
-    case useExternalEmojis = 18
-    case viewGuildInsights = 19
-    case connect = 20
-    case speak = 21
-    case muteMembers = 22
-    case deafenMembers = 23
-    case moveMembers = 24
-    case useVAD = 25
-    case changeNickname = 26
-    case manageNicknames = 27
-    case manageRoles = 28
-    case manageWebhooks = 29
-    case manageGuildExpressions = 30
-    case useApplicationCommands = 31
-    case requestToSpeak = 32
-    case manageEvents = 33
-    case manageThreads = 34
-    case createPublicThreads = 35
-    case createPrivateThreads = 36
-    case useExternalStickers = 37
-    case sendMessagesInThreads = 38
-    case useEmbeddedActivities = 39
-    case moderateMembers = 40
-    case viewCreatorMonetizationAnalytics = 41
-    case useSoundboard = 42
-    case useExternalSounds = 45
-    case sendVoiceMessages = 46
-}
-#endif
 
 /// https://discord.com/developers/docs/topics/permissions#role-object
 public struct Role: Sendable, Codable {
