+++ Sources/DiscordModels/Types/Sticker.swift
@@ -3,21 +3,13 @@
 public struct Sticker: Sendable, Codable {
     
     /// https://discord.com/developers/docs/resources/sticker#sticker-object-sticker-types
-#if swift(>=5.9) && $Macros
     @UnstableEnum<Int>
     public enum Kind: Sendable, Codable {
         case standard // 1
         case guild // 2
     }
-#else
-    public enum Kind: Int, Sendable, Codable, ToleratesIntDecodeMarker {
-        case standard = 1
-        case guild = 2
-    }
-#endif
 
     /// https://discord.com/developers/docs/resources/sticker#sticker-object-sticker-format-types
-#if swift(>=5.9) && $Macros
     @UnstableEnum<Int>
     public enum FormatKind: Sendable, Codable {
         case png // 1
@@ -25,14 +17,6 @@ public struct Sticker: Sendable, Codable {
         case lottie // 3
         case gif // 4
     }
-#else
-    public enum FormatKind: Int, Sendable, Codable, ToleratesIntDecodeMarker {
-        case png = 1
-        case apng = 2
-        case lottie = 3
-        case gif = 4
-    }
-#endif
 
     public var id: StickerSnowflake
     public var pack_id: StickerPackSnowflake?
