+++ Sources/DiscordModels/Types/User.swift
@@ -3,7 +3,6 @@
 public struct DiscordUser: Sendable, Codable {
     
     /// https://discord.com/developers/docs/resources/user#user-object-premium-types
-#if swift(>=5.9) && $Macros
     @UnstableEnum<Int>
     public enum PremiumKind: Sendable, Codable {
         case none // 0
@@ -11,17 +10,8 @@ public struct DiscordUser: Sendable, Codable {
         case nitro // 2
         case nitroBasic // 3
     }
-#else
-    public enum PremiumKind: Int, Sendable, Codable, ToleratesIntDecodeMarker {
-        case none = 0
-        case nitroClassic = 1
-        case nitro = 2
-        case nitroBasic = 3
-    }
-#endif
 
     /// https://discord.com/developers/docs/resources/user#user-object-user-flags
-#if swift(>=5.9) && $Macros
     @UnstableEnum<UInt>
     public enum Flag: Sendable {
         case staff // 0
@@ -40,25 +30,6 @@ public struct DiscordUser: Sendable, Codable {
         case botHttpInteractions // 19
         case activeDeveloper // 22
     }
-#else
-    public enum Flag: UInt, Sendable {
-        case staff = 0
-        case partner = 1
-        case hypeSquad = 2
-        case BugHunterLevel1 = 3
-        case hypeSquadOnlineHouse1 = 6
-        case hypeSquadOnlineHouse2 = 7
-        case hypeSquadOnlineHouse3 = 8
-        case premiumEarlySupporter = 9
-        case teamPseudoUser = 10
-        case bugHunterLevel2 = 14
-        case verifiedBot = 16
-        case verifiedDeveloper = 17
-        case certifiedModerator = 18
-        case botHttpInteractions = 19
-        case activeDeveloper = 22
-    }
-#endif
 
     public var id: UserSnowflake
     public var username: String
@@ -128,7 +99,6 @@ extension DiscordUser {
     public struct Connection: Sendable, Codable {
 
         /// https://discord.com/developers/docs/resources/user#connection-object-services
-#if swift(>=5.9) && $Macros
         @UnstableEnum<String>
         public enum Service: Sendable, Codable {
             case battleNet // "Battle.net"
@@ -151,43 +121,13 @@ extension DiscordUser {
             case xbox // "Xbox"
             case youtube // "YouTube"
         }
-#else
-        public enum Service: String, Sendable, Codable, ToleratesStringDecodeMarker {
-            case battleNet = "Battle.net"
-            case ebay = "eBay"
-            case epicGames = "Epic Games"
-            case facebook = "Facebook"
-            case github = "GitHub"
-            case instagram = "Instagram"
-            case leagueOfLegends = "League of Legends"
-            case paypal = "PayPal"
-            case playstation = "PlayStation Network"
-            case reddit = "Reddit"
-            case riotGames = "Riot Games"
-            case spotify = "Spotify"
-            case skype = "Skype"
-            case steam = "Steam"
-            case tikTok = "TikTok"
-            case twitch = "Twitch"
-            case twitter = "Twitter"
-            case xbox = "Xbox"
-            case youtube = "YouTube"
-        }
-#endif
 
         /// https://discord.com/developers/docs/resources/user#connection-object-visibility-types
-#if swift(>=5.9) && $Macros
         @UnstableEnum<Int>
         public enum VisibilityKind: Sendable, Codable {
             case none // 0
             case everyone // 1
         }
-#else
-        public enum VisibilityKind: Int, Sendable, Codable {
-            case none = 0
-            case everyone = 1
-        }
-#endif
 
         public var id: String
         public var name: String
